- change visibility to be able to share maximum classes with Quarkus langchain - Add OllamaClient interface.
