+++ langchain4j-ollama/src/main/java/dev/langchain4j/model/ollama/EmbeddingRequest.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-class EmbeddingRequest {
+public class EmbeddingRequest {
 
     private String model;
     private String prompt;
