+++ langchain4j-ollama/src/main/java/dev/langchain4j/model/ollama/EmbeddingResponse.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-class EmbeddingResponse {
+public class EmbeddingResponse {
 
     private float[] embedding;
 }
