+++ langchain4j-ollama/src/main/java/dev/langchain4j/model/ollama/OllamaLanguageModel.java
@@ -23,7 +23,7 @@ import static java.time.Duration.ofSeconds;
  */
 public class OllamaLanguageModel implements LanguageModel {
 
-    private final OllamaClient client;
+    private final DefaultOllamaClient client;
     private final String modelName;
     private final Options options;
     private final String format;
@@ -47,7 +47,7 @@ public class OllamaLanguageModel implements LanguageModel {
                                Boolean logResponses,
                                Map<String, String> customHeaders
                                ) {
-        this.client = OllamaClient.builder()
+        this.client = DefaultOllamaClient.builder()
                 .baseUrl(baseUrl)
                 .timeout(getOrDefault(timeout, ofSeconds(60)))
                 .logRequests(logRequests)
