+++ langchain4j-ollama/src/main/java/dev/langchain4j/model/ollama/OllamaModels.java
@@ -11,7 +11,7 @@ import static dev.langchain4j.internal.Utils.getOrDefault;
 
 public class OllamaModels {
 
-    private final OllamaClient client;
+    private final DefaultOllamaClient client;
     private final Integer maxRetries;
 
     @Builder
@@ -21,7 +21,7 @@ public class OllamaModels {
                         Boolean logRequests,
                         Boolean logResponses
                         ) {
-        this.client = OllamaClient.builder()
+        this.client = DefaultOllamaClient.builder()
                 .baseUrl(baseUrl)
                 .timeout((getOrDefault(timeout, Duration.ofSeconds(60))))
                 .logRequests(logRequests)
