+++ langchain4j-ollama/src/main/java/dev/langchain4j/model/ollama/OllamaStreamingChatModel.java
@@ -25,7 +25,7 @@ import static java.time.Duration.ofSeconds;
  */
 public class OllamaStreamingChatModel implements StreamingChatLanguageModel {
 
-    private final OllamaClient client;
+    private final DefaultOllamaClient client;
     private final String modelName;
     private final Options options;
     private final String format;
@@ -47,7 +47,7 @@ public class OllamaStreamingChatModel implements StreamingChatLanguageModel {
                                     Boolean logResponses,
                                     Map<String, String> customHeaders
                                     ) {
-        this.client = OllamaClient.builder()
+        this.client = DefaultOllamaClient.builder()
                 .baseUrl(baseUrl)
                 .timeout(getOrDefault(timeout, ofSeconds(60)))
                 .logRequests(logRequests)
