+++ langchain4j-ollama/src/main/java/dev/langchain4j/model/ollama/Role.java
@@ -1,6 +1,6 @@
 package dev.langchain4j.model.ollama;
 
-enum Role {
+public enum Role {
 
     SYSTEM,
     USER,
