+++ langchain4j-ollama/src/main/java/dev/langchain4j/model/ollama/tool/NoToolsDelegate.java
@@ -1,8 +1,9 @@
-package dev.langchain4j.model.ollama;
+package dev.langchain4j.model.ollama.tool;
 
 import dev.langchain4j.data.message.AiMessage;
 import dev.langchain4j.data.message.ChatMessage;
 import dev.langchain4j.model.chat.ChatLanguageModel;
+import dev.langchain4j.model.ollama.*;
 import dev.langchain4j.model.output.Response;
 import dev.langchain4j.model.output.TokenUsage;
 
@@ -10,14 +11,14 @@ import java.util.List;
 
 import static dev.langchain4j.model.ollama.OllamaMessagesUtils.toOllamaMessages;
 
-class NoToolsDelegate implements ChatLanguageModel {
+public class NoToolsDelegate implements ChatLanguageModel {
 
     private final OllamaClient client;
     private final String modelName;
     private final Options options;
     private final String format;
 
-    NoToolsDelegate(OllamaClient client, String modelName, Options options, String format) {
+    public NoToolsDelegate(OllamaClient client, String modelName, Options options, String format) {
         this.client = client;
         this.modelName = modelName;
         this.options = options;
