+++ langchain4j-ollama/src/test/java/dev/langchain4j/model/ollama/service/AiServicesWithOllamaToolsIT.java
@@ -1,7 +1,7 @@
 package dev.langchain4j.model.ollama.service;
 
 import dev.langchain4j.model.chat.ChatLanguageModel;
-import dev.langchain4j.model.ollama.ExperimentalTools;
+import dev.langchain4j.model.ollama.tool.ExperimentalTools;
 import dev.langchain4j.model.ollama.OllamaChatModel;
 import dev.langchain4j.service.AiServices;
 import org.junit.jupiter.api.Disabled;
